+++ gems/omq-backend-libzmq/lib/omq/backend/libzmq.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 require "omq"
-require_relative "../ffi/libzmq"
-require_relative "../ffi/engine"
+require_relative "libzmq/native"
+require_relative "libzmq/engine"
 
-OMQ::Backend.register(:libzmq, OMQ::FFI::Engine)
-OMQ::Backend.register(:ffi, OMQ::FFI::Engine)
+OMQ::Backend.register(:libzmq, OMQ::Backend::Libzmq::Engine)