+++ test/omq/transport/inproc_test.rb
@@ -7,6 +7,53 @@ describe OMQ::Transport::Inproc do
 
   before { Inproc.reset! }
 
+  it "treats inproc:// as a ruby:// alias" do
+    Async do
+      endpoint = "inproc-alias-#{object_id}"
+      server   = OMQ::PAIR.bind("inproc://#{endpoint}")
+      client   = OMQ::PAIR.connect("inproc://#{endpoint}")
+
+      assert_equal "ruby://#{endpoint}", server.engine.listeners.keys.fetch(0)
+
+      client << "hello"
+      assert_equal ["hello"], server.receive
+    ensure
+      client&.close
+      server&.close
+    end
+  end
+
+
+  it "shares one namespace between ruby:// and inproc://" do
+    Async do
+      endpoint = "inproc-ruby-alias-#{object_id}"
+      server   = OMQ::PAIR.bind("ruby://#{endpoint}")
+      client   = OMQ::PAIR.connect("inproc://#{endpoint}")
+
+      client << "hello"
+      assert_equal ["hello"], server.receive
+    ensure
+      client&.close
+      server&.close
+    end
+  end
+
+
+  it "lets a registered inproc transport override the alias" do
+    original_transports = OMQ::Engine.transports
+    fake_transport      = Module.new
+
+    OMQ::Engine.instance_variable_set(:@transports, original_transports.dup)
+    OMQ::Engine.transports["inproc"] = fake_transport
+
+    engine = OMQ::Engine.new(:PAIR, OMQ::Options.new)
+
+    assert_same fake_transport, engine.transport_for("inproc://native")
+  ensure
+    OMQ::Engine.instance_variable_set(:@transports, original_transports)
+  end
+
+
   describe "Pipe" do
     it "transfers messages bidirectionally" do
       Async do